Se implementa endpoint de GetAll con paginación

+++ Integration.Orchestrator.Backend.Application/Models/ModelResponseGetAll.cs
@@ -2,6 +2,6 @@
 {
     public class ModelResponseGetAll<T> : ModelResponse<T> where T : class
     {
-        public int? TotalRows { get; set; }
+        public long? TotalRows { get; set; }
     }
 }
